Update Partials and Sidebar

+++ routes/web.php
@@ -27,55 +27,63 @@ Route::get('/', function ()
 });
 
 
+// Start Admin Route Groups
+
 Route::middleware(['auth', 'role:admin'])->group(function(){
+
     Route::get('admin/dashboard', [DashboardController::class, 'adminIndex'])->name('admin.dashboard');
     // Start Rute Barang
-    Route::get('/barang', [BarangController::class, 'index'])->name('barang.index');
+    Route::get('/admin/barang', [BarangController::class, 'index'])->name('barang.index');
     // Create Rute Barang
-    Route::get('/barang/tambah', [BarangController::class, 'create'])->name('barang.create');
-    Route::post('/barang', [BarangController::class, 'store'])->name('barang.store');
+    Route::get('/admin/barang/tambah', [BarangController::class, 'create'])->name('barang.create');
+    Route::post('/admin/barang', [BarangController::class, 'store'])->name('barang.store');
     // Edit Rute Barang
-    Route::get('/barang/{id}/edit', [BarangController::class, 'edit'])->name('barang.edit');
-    Route::put('/barang/{id}', [BarangController::class, 'update'])->name('barang.update');
+    Route::get('/admin/barang/{id}/edit', [BarangController::class, 'edit'])->name('barang.edit');
+    Route::put('/admin/barang/{id}', [BarangController::class, 'update'])->name('barang.update');
     // Hapus Rute Barang
-    Route::delete('/barang/{id}', [BarangController::class, 'destroy'])->name('barang.destroy');
+    Route::delete('/admin/barang/{id}', [BarangController::class, 'destroy'])->name('barang.destroy');
     // End Rute Barang
     
     // Start Rute Kategori
-    Route::get('/kategori', [KategoriController::class, 'index'])->name('kategori.index');
+    Route::get('/admin/kategori', [KategoriController::class, 'index'])->name('kategori.index');
     // Create Rute Kategori
-    Route::get('/kategori/tambah', [KategoriController::class, 'create'])->name('kategori.create');
-    Route::post('/kategori', [KategoriController::class, 'store'])->name('kategori.store');
+    Route::get('/admin/kategori/tambah', [KategoriController::class, 'create'])->name('kategori.create');
+    Route::post('/admin/kategori', [KategoriController::class, 'store'])->name('kategori.store');
     // Edit Rute Kategori
-    Route::get('/kategori/{id}/edit', [KategoriController::class, 'edit'])->name('kategori.edit');
-    Route::put('/kategori/{id}', [KategoriController::class, 'update'])->name('kategori.update');
+    Route::get('/admin/kategori/{id}/edit', [KategoriController::class, 'edit'])->name('kategori.edit');
+    Route::put('/admin/kategori/{id}', [KategoriController::class, 'update'])->name('kategori.update');
     // Hapus Rute Kategori
-    Route::delete('/kategori/{id}', [KategoriController::class, 'destroy'])->name('kategori.destroy');
+    Route::delete('/admin/kategori/{id}', [KategoriController::class, 'destroy'])->name('kategori.destroy');
     // End Rute Kategori
     
     // Start Rute Ruangan
-    Route::get('/ruangan', [RuanganController::class, 'index'])->name('ruangan.index');
+    Route::get('/admin/ruangan', [RuanganController::class, 'index'])->name('ruangan.index');
     // Create
-    Route::get('/ruangan/tambah', [RuanganController::class, 'create'])->name('ruangan.create');
-    Route::post('/ruangan', [RuanganController::class, 'store'])->name('ruangan.store');
+    Route::get('/admin/ruangan/tambah', [RuanganController::class, 'create'])->name('ruangan.create');
+    Route::post('/admin/ruangan', [RuanganController::class, 'store'])->name('ruangan.store');
     // Edit
-    Route::get('/ruangan/{id}/edit', [RuanganController::class, 'edit'])->name('ruangan.edit');
-    Route::put('/ruangan/{id}', [RuanganController::class, 'update'])->name('ruangan.update');
+    Route::get('/admin/ruangan/{id}/edit', [RuanganController::class, 'edit'])->name('ruangan.edit');
+    Route::put('/admin/ruangan/{id}', [RuanganController::class, 'update'])->name('ruangan.update');
     // Hapus
-    Route::delete('/ruangan/{id}', [RuanganController::class, 'destroy'])->name('ruangan.destroy');
+    Route::delete('/admin/ruangan/{id}', [RuanganController::class, 'destroy'])->name('ruangan.destroy');
     // End Rute Ruangan
 
     // Start Rute Peminjam
-    Route::get('/peminjam', [PeminjamController::class, 'index'])->name('peminjam.index');
+    Route::get('/admin/peminjam', [PeminjamController::class, 'index'])->name('peminjam.index');
     // Create
-    Route::get('/peminjam/tambah', [PeminjamController::class, 'create'])->name('peminjam.create');
-    Route::post('/peminjam', [PeminjamController::class, 'store'])->name('peminjam.store');
+    Route::get('/admin/peminjam/tambah', [PeminjamController::class, 'create'])->name('peminjam.create');
+    Route::post('/admin/peminjam', [PeminjamController::class, 'store'])->name('peminjam.store');
     // Edit
-    Route::get('/peminjam/{id}/edit', [PeminjamController::class, 'edit'])->name('peminjam.edit');
-    Route::put('/peminjam/{id}', [PeminjamController::class, 'update'])->name('peminjam.update');
+    Route::get('/admin/peminjam/{id}/edit', [PeminjamController::class, 'edit'])->name('peminjam.edit');
+    Route::put('/admin/peminjam/{id}', [PeminjamController::class, 'update'])->name('peminjam.update');
     // Hapus
-    Route::delete('/peminjam/{id}', [PeminjamController::class, 'destroy'])->name('peminjam.destroy');
+    Route::delete('/admin/peminjam/{id}', [PeminjamController::class, 'destroy'])->name('peminjam.destroy');
     // End Rute Peminjam
 
 });    
+// End Admin Route Groups
 
+Route::middleware(['auth', 'role:user'])->group(function()
+{
+    Route::get('dashboard', [DashboardController::class, 'userIndex'])->name('user.dashboard');
+});
